src/test/java/mu/mcb/property/evalution/controller/FacilityControllerTest.java
package mu.mcb.property.evalution.controller;

import static org.mockito.Mockito.when;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;

import java.util.Arrays;
import java.util.List;

import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.mockito.InjectMocks;
import org.mockito.MockitoAnnotations;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.http.MediaType;
import org.springframework.test.web.servlet.MockMvc;

import mu.mcb.property.evalution.dto.FacilityTypeDto;
import mu.mcb.property.evalution.service.FacilityService;

@WebMvcTest(FacilityController.class)
public class FacilityControllerTest {

    @Autowired
    private MockMvc mockMvc;

    @MockBean
    private FacilityService facilityService;

    @InjectMocks
    private FacilityController facilityController;

    @BeforeEach
    void setUp() {
        MockitoAnnotations.openMocks(this);
    }

    @Test
    void testGetFacilityTypes() throws Exception {
        FacilityTypeDto facilityType1 = new FacilityTypeDto(1L, "Type1");
        FacilityTypeDto facilityType2 = new FacilityTypeDto(2L, "Type2");
        List<FacilityTypeDto> facilityTypes = Arrays.asList(facilityType1, facilityType2);

        when(facilityService.getFacilityTypes()).thenReturn(facilityTypes);

        mockMvc.perform(get("/api/v1/facility/facilitytypes")
                .contentType(MediaType.APPLICATION_JSON))
                .andExpect(status().isOk())
                .andExpect(jsonPath("$[0].id").value(1L))
                .andExpect(jsonPath("$[0].name").value("Type1"))
                .andExpect(jsonPath("$[1].id").value(2L))
                .andExpect(jsonPath("$[1].name").value("Type2"));
    }
}
